requirements.txt
langchain>=0.1.16
langchain-core>=0.1.35
langchain-google-genai>=0.0.6
langchain-text-splitters
langchain-openai
langchain-community
google-generativeai
openai
aiofiles
tree_sitter
tiktoken
gitpython
pytest
pytest-asyncio